requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pymongo==4.10.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
groq==0.13.0
requests==2.32.3
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355
python-multipart==0.0.12
PyJWT==2.9.0
firebase-admin==6.8.0
protobuf==5.29.6
grpcio-status==1.67.1
APScheduler==3.10.4
pytest==8.3.3
pdfplumber==0.11.5
pikepdf==9.4.2
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
rapidocr-onnxruntime==1.3.24
streamlit==1.38.0
